core/serializers/tracking_serializers.py
from rest_framework import serializers
from core.models.tracking_models import Checklist


class ChecklistSerializer(serializers.ModelSerializer):
    class Meta:
        model = Checklist
        fields = "__all__"
